Update project files
The packages of the julia ecosystem are moving away from the hacked
system that still required the old REQUIRE files.

This commit removes the code of conduct, contributing and humans files.

It also replaces the REQUIRE file with a Project.toml file, and updates
the readme file.

A coverage folder is added along with it's own project TOML file.

The appveyor configuration file is also updated to match the more recent
julia package examples.