Project update (#33)
This removes testing for julia v0.6 and removes the use of Nullables.

Prior to julia 1.0 (and 0.7), the tryparse method would return
Nullable{T} values.

But now they return either an AminoAcid or "nothing",
since improvements in julia v1.0 and v0.7 mean simple
unions are now actually efficient.

DNA and RNA is unified under an abstract `NucleicAcid` type, which
allows us to rewrite some methods more generically.

The print and show methods have been re-written into generic forms,
as have the `isless` and other operators.

The packages of the julia ecosystem are moving away from a hacked
system that still required the old REQUIRE files.

This commit removes the code of conduct, contributing and humans files.

It also replaces the REQUIRE file with a Project.toml file, and updates
the readme file.

A coverage folder is added along with its own project TOML file.

The appveyor configuration file is also updated to match the more recent
julia package examples.